requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
selenium==4.15.2
playwright==1.40.0
groq==0.4.2
python-dotenv==1.0.0
xhtml2pdf==0.2.11
tqdm==4.66.1
webdriver-manager==4.0.1
requests==2.31.0
gunicorn==21.2.0
Pillow==10.0.0